Backend/requirements.txt
# ============================================================================
# Sentinel Core + Master Agent Forensics System Requirements
# ============================================================================

# ============================================================================
# Deep Learning & Neural Networks
# ============================================================================
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.12.0
keras>=2.12.0

# ============================================================================
# Computer Vision & Object Detection
# ============================================================================
opencv-python>=4.8.0
ultralytics>=8.0.0  # YOLOv8
Pillow>=9.5.0
scikit-image>=0.20.0

# ============================================================================
# Natural Language Processing & Text
# ============================================================================
transformers>=4.30.0  # HuggingFace models (BERT, RoBERTa, DistilBERT)
spacy>=3.5.0
nltk>=3.8.0
pytesseract>=0.3.10  # OCR
pdf2image>=1.16.0
PyPDF2>=3.0.0
python-pptx>=0.6.21

# ============================================================================
# Audio Processing
# ============================================================================
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.10.0

# ============================================================================
# Large Language Models
# ============================================================================
google-generativeai>=0.3.0  # Google Gemini API for summarization
llama-cpp-python>=0.2.0  # Local LLaMA support

# ============================================================================
# Specialized ML Models
# ============================================================================
scikit-learn>=1.3.0  # TabNet, decision trees, malware detection
xgboost>=2.0.0
catboost>=1.2.0
TabNet  # TabNet for anomaly detection
pefile>=2022.8.18  # PE header parsing for malware analysis
joblib>=1.3.0  # Model serialization for malware classifiers

# ============================================================================
# Vision-Text Multimodal
# ============================================================================
#clip-by-openai>=1.0.0

# ============================================================================
# Data Processing & Utilities
# ============================================================================
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
tqdm>=4.65.0  # Progress bars

# File type detection (optional - system will fallback to extension-based detection)
# Windows users: Use python-magic-bin which includes libmagic DLLs
# python-magic-bin>=0.4.14; sys_platform == 'win32'
# python-magic>=0.4.27; sys_platform != 'win32'
yara-python>=4.5.0  # Optional YARA rule scanning for file integrity

# ============================================================================
# Database & Serialization
# ============================================================================
# sqlite3 is included in Python stdlib

# ============================================================================
# Logging & Monitoring
# ============================================================================
python-dotenv>=1.0.0
psutil>=5.9.0  # System monitoring
GPUtil>=1.4.0  # GPU monitoring

# ============================================================================
# Web & API
# ============================================================================
requests>=2.31.0
aiohttp>=3.8.0
fastapi>=0.110.0
uvicorn>=0.29.0
reportlab>=4.0.9

# ============================================================================
# Configuration & Type Hints
# ============================================================================
pydantic>=2.0.0
typing-extensions>=4.7.0

# ============================================================================
# Web API & FastAPI
# ============================================================================
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# ============================================================================
# Database & ORM
# ============================================================================
sqlalchemy>=2.0.0

# ============================================================================
# Development & Quality
# ============================================================================
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.0

# ============================================================================
# Optional: For Jupyter Notebook Support
# ============================================================================
# jupyter>=1.0.0
# jupyterlab>=4.0.0
# ipython>=8.0.0
